+++ pimdb/src/indexes/sorted.bench.ts
