+++ pimdb/src/indexes/substring.bench.ts
