+++ pimdb/src/indexes/substring.test.ts
