+++ pimdb/src/indexes/substring.ts
